Remove listening to host connected/disconnected from custom app to core activity, remove unused methods from custom app, changed the function scope of getPlaylistService to private, cleaned up imports, added some comments
